Another compilation fix

+++ src/fheroes2/game/game_interface.cpp
@@ -265,7 +265,8 @@ int32_t Interface::Basic::GetDimensionDoorDestination( const int32_t from, const
                 valid = ( ( spellROI & mp ) && MP2::isClearGroundObject( tile.GetObject() ) && water == world.GetTiles( dst ).isWater() );
             }
 
-            cursor.SetThemes( valid ? ( water ? Cursor::CURSOR_HERO_BOAT : Cursor::CURSOR_HERO_MOVE ) : Cursor::WAR_NONE );
+            cursor.SetThemes( valid ? ( water ? static_cast<int>( Cursor::CURSOR_HERO_BOAT ) : static_cast<int>( Cursor::CURSOR_HERO_MOVE ) )
+                              : static_cast<int>( Cursor::WAR_NONE ) );
 
             if ( dst >= 0 && le.MousePressRight() ) {
                 const Maps::Tiles & tile = world.GetTiles( dst );
